Use Minitest::StatisticsReporter
Minitest::StatisticsReporter#report already gathers the statistics, so
we can delegate stat-gathering to it and just call our notifier with the
stats.
