+++ lib/minitest/guard_minitest_plugin.rb
@@ -1,4 +1,8 @@
-require 'guard/minitest/reporter'
+if ::MiniTest::Unit::VERSION =~ /^5/
+  require 'guard/minitest/reporter'
+else
+  require 'guard/minitest/reporters/old_reporter'
+end
 
 module Minitest
   def self.plugin_guard_minitest_options(opts, options) # :nodoc:
